+++ core/lib.c
@@ -139,7 +139,7 @@ size_t unistrlower (char *dst, const char *src, size_t ds)
 	    sout++;
 	    ds--;
 	  }
-	  if (len == -(ssize_t)2) /* premature end of string */
+	  if (len == -2) /* premature end of string */
 	    break;
 	  ss--;
 	  ch++; /* and skip bad char */
@@ -562,7 +562,7 @@ static ssize_t _try_subst (char *buf, size_t bs, const char *text, size_t s)
   size_t n = safe_strlen (text);
 
   if (!n || s == 0)			/* don't print, don't wrap */
-    return -(ssize_t)1;
+    return -1;
   if (n > s && bs == 0)			/* wrap to next line */
     return 0;
   else if (n > s && n > bs)		/* no so much space available */
@@ -775,7 +775,7 @@ static char *_try_printl (char *buf, size_t s, printl_t *p, size_t ll, int q)
 	  break;
 	case 'P':			/* port number, zero is "" */
 	  snprintf (tbuf, sizeof(tbuf), "%.0hu", p->port);
-	  n = -(ssize_t)1;		/* in case of zero mark as empty */
+	  n = -1;			/* in case of zero mark as empty */
 	  break;
 	case 't':			/* current time */
 	  n = _try_subst (c, nmax, TimeString, nn);
